+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/feature/Misc.java
@@ -1,0 +1,33 @@
+package insane96mcp.iguanatweaksreborn.module.misc.feature;
+
+import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.insanelib.base.Feature;
+import insane96mcp.insanelib.base.Label;
+import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
+import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+
+@Label(name = "Misc", description = "Various stuff that doesn't fit in any other Feature.")
+@LoadFeature(module = Modules.Ids.MISC)
+public class Misc extends Feature {
+
+    @Config
+    @Label(name = "Prevent fire with resistance", description = "If true, entities will no longer be set on fire if has Fire Resistance (like bedrock edition)")
+    public static Boolean preventFireWithResistance = true;
+
+    public Misc(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+        super(module, enabledByDefault, canBeDisabled);
+    }
+
+    public static boolean isFireImmune(Entity entity) {
+        if (!isEnabled(Misc.class)
+                || !preventFireWithResistance
+                || !(entity instanceof LivingEntity livingEntity))
+            return false;
+
+       return livingEntity.hasEffect(MobEffects.FIRE_RESISTANCE);
+    }
+}